fix: [YTreeView] change updateActive to true twice, activeSet reversed on init
