+++ packages/yuyeon/package.json
@@ -1,6 +1,6 @@
 {
   "name": "yuyeon",
-  "version": "0.3.0-rc.5",
+  "version": "0.3.0-rc.6",
   "keywords": [
     "UI Library",
     "Vue"
